requirements.txt
fastapi
uvicorn
python-multipart
requests
symspellpy
num2words
transformers
torch
pypdf
python-docx
pandas